sea_query

+++ src/service/goods_service.rs
@@ -5,9 +5,9 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Deserialize, Serialize)]
 struct ListQuery {
     #[serde(rename = "pageNumber")]
-    page_number: usize,
+    page_number: u64,
     #[serde(rename = "pageSize")]
-    page_size: usize,
+    page_size: u64,
     name: Option<String>,
 }
